back-end/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
databases
python-dotenv
psycopg2-binary
alembic